Rename: FolderData 컴포넌트의 이름을 FolderCardCollection로 변경
