+++ src/components/Folder.js
@@ -1,16 +1,16 @@
 import HeadNav from "components/HeadNav.js";
 import Footer from "components/Footer.js";
 import FolderProFile from "components/FolderProfile";
-import FolderData from "components/FolderData";
+import FolderData from "components/FolderCardCollection";
 
 function Folder() {
 	return (
-		<div className="folder">
+		<>
 			<HeadNav />
 			<FolderProFile />
 			<FolderData />
 			<Footer />
-		</div>
+		</>
 	);
 }
